Initial Commit
Adding some starter code for a simple stock server + client example. The
plan is to spawn up multiple clients (with some randomization to their
initial capital) and ultimately have multiple threads randomly placing
orders for each client. The hope is to create many unique race
conditions that have to be solved for.